+++ src/store/actions/actions.js
@@ -1,4 +1,4 @@
-import types from './types';
+import types from '../types';
 
 function updateCompose(compose) {
     return {
