+++ src/store/reducers/composes.js
@@ -1,4 +1,4 @@
-import types from './types';
+import types from '../types';
 
 // Example of action.compose
 // {
